Added IOMacros.h, which contains a list of macros to make Workman's IO faster to implement.  Included are macros to get the open and close token of a class, Extract, Insert, and operators << and >> named OUT_OPERATOR and IN_OPERATOR respectively.
Added Hand, mostly as an example of proper use of IOMacros.h, and to get 
started on adding base classes.

Added Chip, for convenience.  It is just a typedef of an unsigned, which 
can be converted to a class later if needed.

Took approximately 45 minutes to make the classes and macros, 15 minutes 
to verify.